bugs fucked up me everytime still not solved just for streak

+++ src/main/java/com/example/spring_api_starter/dtos/ProductDto.java
@@ -17,5 +17,5 @@ public class ProductDto {
     private String name;
     private String description;
     private BigDecimal price;
-    private byte categoryId;
+    private long categoryId;
 }
